+++ numba_utils/__init__.py
@@ -105,7 +105,7 @@ from numba_utils.random import (
     weighted_sampling,
 )
 
-__version__ = "0.4.3"
+__version__ = "0.4.4"
 
 __all__ = [
     "BenchmarkResult",